requirements.txt
aiogram==2.22.1
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
beautifulsoup4==4.11.1
certifi==2022.9.14
chardet==3.0.4
charset-normalizer==2.1.1
frozenlist==1.3.1
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.9.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
langid==1.1.6
multidict==6.0.2
numpy==1.23.3
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
urllib3==1.26.12
yarl==1.8.1